feat: add project mapper with policy-based approach and schema detection
